requirements.txt
flask==2.3.3
flask-limiter==3.5.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.54.3
anthropic==0.7.7
redis==5.0.1
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0